Server/Port/Cluster mbeans
git-svn-id: svn://svn.caucho.com/resin/trunk@1080 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/server/webapp/StatisticsFilterChain.java
@@ -69,8 +69,8 @@ public class StatisticsFilterChain
 
         long time = Alarm.getExactTime();
 
-        int readBytes = -1;
-        int writeBytes = -1;
+        long readBytes = -1;
+        long writeBytes = -1;
 
         readBytes = tcpConnection.getSocket().getTotalReadBytes();
         writeBytes = tcpConnection.getSocket().getTotalWriteBytes();
@@ -89,7 +89,7 @@ public class StatisticsFilterChain
         readBytes = tcpConnection.getSocket().getTotalReadBytes() - readBytes;
         writeBytes = tcpConnection.getSocket().getTotalReadBytes() - writeBytes;
 
-        _application.updateStatistics(time, readBytes, writeBytes, clientDisconnectException != null);
+        _application.updateStatistics(time, (int) readBytes, (int) writeBytes, clientDisconnectException != null);
 
         if (clientDisconnectException != null)
           throw clientDisconnectException;
